requirements.txt
Django==3.1
djangorestframework
